Fixes FILL_WITH_MEAN missing value strategy with appropriate cast
