+++ tests/integration_tests/test_missing_value_strategy.py
@@ -18,9 +18,10 @@ import tempfile
 
 import numpy as np
 import pandas as pd
+import pytest
 
 from ludwig.api import LudwigModel
-from ludwig.constants import DROP_ROW, PREPROCESSING, TRAINER
+from ludwig.constants import DROP_ROW, FILL_WITH_MEAN, PREPROCESSING, TRAINER
 from tests.integration_tests.utils import (
     binary_feature,
     category_feature,
@@ -32,6 +33,7 @@ from tests.integration_tests.utils import (
     set_feature,
     text_feature,
     vector_feature,
+    init_backend,
 )
 
 
@@ -63,6 +65,27 @@ def test_missing_value_prediction(csv_filename):
 
         model = LudwigModel.load(os.path.join(output_dir, "model"))
         model.predict(dataset=dataset)
+
+
+@pytest.mark.parametrize("backend", ["local", "ray"])
+@pytest.mark.distributed
+def test_missing_values_fill_with_mean(backend, csv_filename, tmpdir):
+    data_csv_path = os.path.join(tmpdir, csv_filename)
+
+    kwargs = {PREPROCESSING: {"missing_value_strategy": FILL_WITH_MEAN}}
+    input_features = [
+        number_feature(**kwargs),
+        binary_feature(),
+        category_feature(vocab_size=3),
+    ]
+    output_features = [binary_feature()]
+    training_data_csv_path = generate_data(input_features, output_features, data_csv_path)
+
+    config = {"input_features": input_features, "output_features": output_features, TRAINER: {"epochs": 2}}
+    with init_backend(backend):
+        # run preprocessing
+        ludwig_model = LudwigModel(config, backend=backend)
+        ludwig_model.preprocess(dataset=training_data_csv_path)
 
 
 def test_missing_values_drop_rows(csv_filename, tmpdir):
