+++ tests/integration_tests/test_preprocessing.py
@@ -1,4 +1,3 @@
-import contextlib
 import os
 
 import numpy as np
@@ -12,24 +11,9 @@ from tests.integration_tests.utils import (
     category_feature,
     generate_data,
     LocalTestBackend,
-    ray_cluster,
     sequence_feature,
+    init_backend,
 )
-
-
-@contextlib.contextmanager
-def init_backend(backend: str):
-    if backend == "local":
-        with contextlib.nullcontext():
-            yield
-            return
-
-    if backend == "ray":
-        with ray_cluster():
-            yield
-            return
-
-    raise ValueError(f"Unrecognized backend: {backend}")
 
 
 @pytest.mark.parametrize("backend", ["local", "ray"])
